fix(ocr): Calling tesseract in subprocess
`click_point = ocr_algo.find_text_in_image()` returned a Deferred for
the click point, which was never waited for. This resulted in
> INFO [root:cli:run_on_img]: Final click point: <DeferredList at 0xADDR>

In some cases this stopped the reactor to early, leading to the main
process shutting down and closing the pipes to Tesseract, which then
terminated and return nothing.

As we're not interested in STDOUT/STDERR of Tesseract just use
`twisted.internet.utils.getProcessValue()` to wait for Tesseract to
teminate successfully.

Make sure to stop the reactor in all cases.